requirements.txt
google-generativeai>=0.8.0
requests>=2.28.0
python-dotenv>=1.0.0
